Move MarketHourAwareConsolidator to Data.Consolidators namespace
