+++ Algorithm.CSharp/MarketHourAwareIntradayConsolidationRegressionAlgorithm.cs
@@ -16,7 +16,7 @@
 using System;
 using System.Collections.Generic;
 using QuantConnect.Data;
-using QuantConnect.Data.Common;
+using QuantConnect.Data.Consolidators;
 using QuantConnect.Data.Market;
 using QuantConnect.Interfaces;
 using QuantConnect.Securities;
