+++ Algorithm.CSharp/SecuritySessionWithFutureContractRegressionAlgorithm.cs
@@ -16,7 +16,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using QuantConnect.Data;
-using QuantConnect.Data.Common;
+using QuantConnect.Data.Consolidators;
 using QuantConnect.Data.Market;
 using QuantConnect.Securities;
 using QuantConnect.Securities.Future;
