+++ Engine/DataFeeds/AggregationManager.cs
@@ -14,7 +14,6 @@
 */
 
 using QuantConnect.Data;
-using QuantConnect.Data.Common;
 using QuantConnect.Data.Consolidators;
 using QuantConnect.Data.Market;
 using QuantConnect.Lean.Engine.DataFeeds.Enumerators;
